Updated for changes and poster image
1. The loading bar is the correct color but I forgot to design the
background color below the loading bar. It should be #424242.
2. Player should be 30px tall, currently there are some pixels missing
off the bottom so it�s uneven
3. Play and Pause icons look smashed/stretched. Let me know if I need
to provide other assets besides the original illustrator file.
4. The hover state should be per icon, so the lighter gray icons should
only show individually
5. The large play icon in the center has a few things: missing solid
white part, should not have a glow/shadow, missing shadow overlay on
video
6. The large play icon should not be displayed when the video is
paused, it is only for the first time the viewer sees the video (if
auto-play is turned off)
7. The player should not overlap the video (this was mentioned below -
�b� option is better)
8. The hover state for the circle button
9. The status bar at the bottom is showing up on top of the Video, so
the bottom part of the video is getting cut-off. We need to either (a)
show the bottom bar as an overlay or (b) shift the video up so it is
shown fully.